+++ packages/nestjs-nats-jetstream-transport/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./server";
 export * from "./client";
 export * from "./nats-jetstream-transport.module";
-export * from "./interfaces";
-export * from "./utils/server-consumer-options-builder";
+// export * from "./interfaces/nats-jetstream-client-options.interface";
+// export * from "./interfaces/nats-jetstream-server-options.interface";
+// export * from "./interfaces/server-consumer-options.interface"
+// export * from "./utils/server-consumer-options-builder";
 export * from "./nats-jetstream.context";
