+++ packages/nestjs-nats-jetstream-transport/src/interfaces/nats-jetstream-server-options.interface.ts
@@ -1,0 +1,10 @@
+import { ConnectionOptions, JetStreamOptions } from "nats";
+import { ServerConsumerOptions } from "./server-consumer-options.interface";
+
+export interface NatsJetStreamServerOptions {
+  id: string;
+  connectionOptions: ConnectionOptions;
+  consumerOptions: Partial<ServerConsumerOptions>;
+  jetStreamOptions?: JetStreamOptions;
+}
+
