+++ packages/nestjs-nats-jetstream-transport/src/interfaces/server-consumer-options.interface.ts
@@ -1,31 +1,3 @@
-import { ModuleMetadata } from "@nestjs/common";
-import {
-  ConnectionOptions,
-  JetStreamOptions,
-  JetStreamPublishOptions,
-} from "nats";
-
-export interface NatsJetStreamClientOptions {
-  connectionOptions: ConnectionOptions;
-  jetStreamOption?: JetStreamOptions;
-  jetStreamPublishOptions?: JetStreamPublishOptions;
-}
-
-export interface NatsJetStreamClientAsyncOptions
-  extends Pick<ModuleMetadata, "imports"> {
-  useFactory: (
-    ...args: any[]
-  ) => Promise<NatsJetStreamClientOptions> | NatsJetStreamClientOptions;
-  inject?: any[];
-}
-
-export interface NatsJetStreamServerOptions {
-  id: string;
-  connectionOptions: ConnectionOptions;
-  consumerOptions: Partial<ServerConsumerOptions>;
-  jetStreamOptions?: JetStreamOptions;
-}
-
 export interface ServerConsumerOptions {
   // https://nats.io/blog/jetstream-java-client-03-consume/
   durable?: boolean;
